// TRC-Beneficios-Back/src/interfaces/IBeneficios.ts
export interface beneficiariosAttributes {
  cpf: number;
  nome: string;
  data_nascimento: string;
  genero?: string;
  rua: string;
  telefone?: number;
  celular: number;
  rg?: string;
  data_obito?: string;
  email: string;
  c_documentos?: string;
  representante_legal?: string;
  cep: string;
  bairro: string;
  cidade: string;
  num_complemento: string;
  deficiencia?: string;
}